Move command releated classes into the command folder
